Create Modded YAMS server End Of Stream??
I have a YAMS server, and i have set the server to CraftBukit, and i have put the CraftBukkit.jar into lib in the YAMS folder, and i put the exctracted ZIP file of the mod (Explosives + Mod) into the server. But when i go into the game, go into creative and try to give myself a block from the mod, it crashes. did i do something wrong when installing it? do i need ForgeModLoader in the server? if so, how? thanks for your time!